[betaFam-minor-3]
1. separated `SR` and `SS` tests for `test_gamma_d.py` into `test_gamma_sr_d.py` and `test_gamma_ss_d.py`
2. minor changes in `test_exponential_d.py`
